Update UI to remove row upon deletion
