+++ frontend/src/components/contacts-table-components/contacts-table.jsx
@@ -1,8 +1,9 @@
-import { useEffect, useState } from "react";
+import { useEffect, useState , useContext} from "react";
 import { supabase } from "../../providers/supabaseClient";
 import { UserAuth } from "../../context/AuthContext";
 import { columns, Contact } from "./columns";
 import DataTable from "./data-table";
+import { toast } from "sonner";
 
 const ContactsTable = () => {
   const { session } = UserAuth();
@@ -26,10 +27,29 @@ const ContactsTable = () => {
     fetchContacts();
   }, [session])
 
+  const onDeleteContact = (id) => {
+    console.log(id)
+
+    const deleteContact = async (id) => {
+      const {data, error} = await supabase.from("connections").delete().eq("id", id)
+
+      if (error) {
+        toast.error("There was an error deleting this contact")
+      } else {
+        setData(prevData => prevData.filter(contact => contact.id !== id));
+        toast.success("Contact successfully deleted ✅")
+        
+      }
+    }
+    
+    deleteContact(id)
+  }
+
+
   return (
-    <div className="mx-auto my-10 ">
-      <DataTable columns={columns} data={data} />
-    </div>
+      <div className="mx-auto my-10 ">
+        <DataTable columns={columns(onDeleteContact)} data={data}/>
+      </div>
   );
 }
